+++ resources/views/components/listing.blade.php
@@ -3,7 +3,7 @@
 @pushOnce('head', 'es_url-preconnect')
     <link rel="preconnect" href="{{ config('rapidez.es_url') }}">
 
-    @if($file = vite_filename_path('Listing.vue'))
+    @if ($file = vite_filename_path('Listing.vue'))
         @vite([$file])
     @endif
 @endPushOnce