+++ src/Models/ProductLink.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Rapidez\Core\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ProductLink extends Model
+{
+    const CREATED_AT = null;
+    const UPDATED_AT = null;
+
+    protected $table = 'catalog_product_super_link';
+
+    protected $primaryKey = 'link_id';
+
+    protected $guarded = [];
+}