+++ src/Models/Scopes/Product/WithProductChildrenScope.php
@@ -16,7 +16,7 @@ class WithProductChildrenScope implements Scope
         $attributeModel = config('rapidez.models.attribute');
 
         $superAttributes = Arr::pluck($attributeModel::getCachedWhere(function ($attribute) {
-            return $attribute['super'];
+            return $attribute['super'] && $attribute['flat'];
         }), 'code');
 
         $grammar = $builder->getQuery()->getGrammar();
@@ -30,6 +30,7 @@ class WithProductChildrenScope implements Scope
         $builder
             ->selectRaw('JSON_REMOVE(JSON_OBJECTAGG(IFNULL(children.entity_id, "null__"), JSON_OBJECT(
                 ' . Eventy::filter('product.children.select', <<<QUERY
+                    "sku", children.sku,
                     "price", children.price,
                     "special_price", children.special_price,
                     "special_from_date", DATE(children.special_from_date),