+++ src/Models/Scopes/Product/WithProductSuperAttributesScope.php
@@ -14,7 +14,7 @@ class WithProductSuperAttributesScope implements Scope
     {
         $attributeModel = config('rapidez.models.attribute');
         $superAttributes = Arr::pluck($attributeModel::getCachedWhere(function ($attribute) {
-            return $attribute['super'];
+            return $attribute['super'] && $attribute['flat'];
         }), 'code', 'id');
 
         $grammar = $builder->getQuery()->getGrammar();